Refactor Analysis components
* Move AnalysisServicePage.jsx and AnalysisServiceList.jsx into a AnalysisServicePage directory and document them
* Move AnalysisForm.jsx and AnalysisList.jsx into the source directory
